Splitting the plotting into subclasses that handle binary and HDF5.
